+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * refactor code (sync_scroll_library)
 * breaking change: SyncPageController to LinkPageController
+* add getHeight and getWidth for [CellStyle] to override height and width base on row/column
 
 ## 3.0.1